# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi TEST Project Documentation",
        "description": "Bortsov Ilia 2022",
        "contact": {
            "email": "ibortsov-dev@yandex.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/drive-list": {
            "get": {
                "tags": [
                    "Drive"
                ],
                "summary": "Driving list",
                "operationId": "30716e0eb8099d31c4f5e0000f164840",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/drive": {
            "post": {
                "tags": [
                    "Drive"
                ],
                "summary": "Start driving vehicle by user",
                "operationId": "3b3309059ccbf9318fbed167262a336b",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "user_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vehicle_id",
                        "in": "query",
                        "description": "vehicle_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "user or vehicle not found"
                    },
                    "406": {
                        "description": "user or vehicle are already in used"
                    }
                }
            }
        },
        "/stop": {
            "post": {
                "tags": [
                    "Drive"
                ],
                "summary": "Srop driving vehicle by user",
                "operationId": "780a96c35bde00ce51ae72ee436c4113",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "user_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vehicle_id",
                        "in": "query",
                        "description": "vehicle_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, driving stop"
                    },
                    "404": {
                        "description": "no active driving"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "da45186ba1143d644673983284dd8b24",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/vehicle": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Get list of Vehicle",
                "description": "Returns list of Vehicle",
                "operationId": "23df08b95fdc5e2dbe4e7880c2d2c2b2",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    }
}